Address changes by mtobvin
